room: port Room_TestSectorTrigger to TRX
This unifies sector trigger testing in TRX.
