+++ src/tr1/game/room.c
@@ -1,67 +1,6 @@
 #include "game/room.h"
 
 #include "game/box.h"
-#include "game/camera.h"
-#include "game/items.h"
-#include "game/lara/misc.h"
-#include "game/lot.h"
-#include "game/music.h"
-#include "game/objects/common.h"
-#include "game/savegame.h"
-#include "game/shell.h"
-#include "game/sound.h"
-#include "game/stats.h"
-#include "global/vars.h"
-
-#include <libtrx/game/game.h>
-#include <libtrx/game/game_buf.h>
-#include <libtrx/game/gym.h>
-#include <libtrx/game/objects/general/keyhole.h>
-#include <libtrx/game/objects/general/pickup.h>
-#include <libtrx/game/objects/general/switch.h>
-#include <libtrx/utils.h>
-
-static void M_TriggerMusicTrack(int16_t track, const TRIGGER *const trigger);
-
-static bool M_TestLava(const ITEM *const item);
-
-static void M_TriggerMusicTrack(int16_t track, const TRIGGER *const trigger)
-{
-    if (track == MX_UNUSED_0
-        && (trigger->type == TT_ANTIPAD || trigger->type == TT_ANTITRIGGER)) {
-        Music_Stop();
-        return;
-    }
-
-    if (track <= MX_UNUSED_1 || track >= MAX_MUSIC_TRACKS
-        || !Gym_CanPlayMusicTrack(&track)) {
-        return;
-    }
-
-    uint16_t flags = Music_GetTrackFlags(track);
-    if (flags & IF_ONE_SHOT) {
-        return;
-    }
-
-    if (trigger->type == TT_SWITCH) {
-        flags ^= trigger->mask;
-    } else if (trigger->type == TT_ANTIPAD || trigger->type == TT_ANTITRIGGER) {
-        flags &= -1 - trigger->mask;
-    } else if (trigger->mask) {
-        flags |= trigger->mask;
-    }
-
-    if ((flags & IF_CODE_BITS) == IF_CODE_BITS) {
-        if (trigger->one_shot) {
-            flags |= IF_ONE_SHOT;
-        }
-        Music_Play(track, MPM_TRACKED);
-    } else {
-        Music_StopTrack(track);
-    }
-
-    Music_SetTrackFlags(track, flags);
-}
 
 int16_t Room_GetTiltType(const SECTOR *sector, int32_t x, int32_t y, int32_t z)
 {
@@ -285,300 +224,5 @@ void Room_AlterFloorHeight(const ITEM *const item, const int32_t height)
         } else {
             box->overlap_index &= ~BOX_BLOCKED;
         }
-    }
-}
-
-static bool M_TestLava(const ITEM *const item)
-{
-    if (item->hit_points < 0 || g_Lara.water_status == LWS_CHEAT
-        || (g_Lara.water_status == LWS_ABOVE_WATER
-            && item->pos.y != item->floor)) {
-        return false;
-    }
-
-    // OG fix: check if floor index has lava
-    int16_t room_num = item->room_num;
-    const SECTOR *const sector =
-        Room_GetSector(item->pos.x, MAX_HEIGHT, item->pos.z, &room_num);
-    return sector->is_death_sector;
-}
-
-void Room_TestSectorTrigger(const ITEM *const item, const SECTOR *const sector)
-{
-    const bool is_heavy = item->object_id != O_LARA;
-    if (!is_heavy && sector->is_death_sector && M_TestLava(item)) {
-        Lara_TouchLava();
-    }
-
-    const TRIGGER *const trigger = sector->trigger;
-    if (trigger == nullptr) {
-        return;
-    }
-
-    if (g_Camera.type != CAM_HEAVY) {
-        Camera_RefreshFromTrigger(trigger);
-    }
-
-    bool switch_off = false;
-    bool flip_map = false;
-    bool flip_available = false;
-    int32_t new_effect = -1;
-    ITEM *camera_item = nullptr;
-    const bool flip_status = Room_GetFlipStatus();
-
-    if (is_heavy) {
-        if (trigger->type != TT_HEAVY) {
-            return;
-        }
-    } else {
-        switch (trigger->type) {
-        case TT_SWITCH: {
-            if (!Switch_Trigger(trigger->item_index, trigger->timer)) {
-                return;
-            }
-            switch_off =
-                Item_Get(trigger->item_index)->current_anim_state == LS_RUN;
-            break;
-        }
-
-        case TT_PAD:
-        case TT_ANTIPAD:
-            if (item->pos.y != item->floor) {
-                return;
-            }
-            break;
-
-        case TT_KEY: {
-            if (!Keyhole_Trigger(trigger->item_index)) {
-                return;
-            }
-            break;
-        }
-
-        case TT_PICKUP: {
-            if (!Pickup_Trigger(trigger->item_index)) {
-                return;
-            }
-            break;
-        }
-
-        case TT_HEAVY:
-        case TT_DUMMY:
-            return;
-
-        case TT_COMBAT:
-            if (g_Lara.gun_status != LGS_READY) {
-                return;
-            }
-            break;
-
-        default:
-            break;
-        }
-    }
-
-    const TRIGGER_CMD *cmd = trigger->command;
-    for (; cmd != nullptr; cmd = cmd->next_cmd) {
-        switch (cmd->type) {
-        case TO_OBJECT: {
-            const int16_t item_num = (int16_t)(intptr_t)cmd->parameter;
-            ITEM *const item = Item_Get(item_num);
-            if (item->flags & IF_ONE_SHOT) {
-                break;
-            }
-
-            item->timer = trigger->timer;
-            if (item->timer != 1) {
-                item->timer *= LOGIC_FPS;
-            }
-
-            if (trigger->type == TT_SWITCH) {
-                item->flags ^= trigger->mask;
-            } else if (
-                trigger->type == TT_ANTIPAD
-                || trigger->type == TT_ANTITRIGGER) {
-                item->flags &= -1 - trigger->mask;
-            } else if (trigger->mask) {
-                item->flags |= trigger->mask;
-            }
-
-            if ((item->flags & IF_CODE_BITS) != IF_CODE_BITS) {
-                break;
-            }
-
-            if (trigger->one_shot) {
-                item->flags |= IF_ONE_SHOT;
-            }
-
-            if (!item->active) {
-                const OBJECT *const obj = Object_Get(item->object_id);
-                if (obj->activate_func != nullptr) {
-                    obj->activate_func(item);
-                } else if (obj->intelligent) {
-                    if (item->status == IS_INACTIVE) {
-                        item->touch_bits = 0;
-                        item->status = IS_ACTIVE;
-                        Item_AddActive(item_num);
-                        LOT_EnableBaddieAI(item_num, 1);
-                    } else if (item->status == IS_INVISIBLE) {
-                        item->touch_bits = 0;
-                        if (LOT_EnableBaddieAI(item_num, 0)) {
-                            item->status = IS_ACTIVE;
-                        } else {
-                            item->status = IS_INVISIBLE;
-                        }
-                        Item_AddActive(item_num);
-                    }
-                } else {
-                    item->touch_bits = 0;
-                    item->status = IS_ACTIVE;
-                    Item_AddActive(item_num);
-                }
-            }
-            break;
-        }
-
-        case TO_CAMERA: {
-            const TRIGGER_CAMERA_DATA *const cam_data =
-                (TRIGGER_CAMERA_DATA *)cmd->parameter;
-            OBJECT_VECTOR *const camera =
-                Camera_GetFixedObject(cam_data->camera_num);
-            if (camera->flags & IF_ONE_SHOT) {
-                break;
-            }
-
-            g_Camera.num = cam_data->camera_num;
-
-            if (g_Camera.type == CAM_LOOK || g_Camera.type == CAM_COMBAT) {
-                break;
-            }
-
-            if (trigger->type == TT_COMBAT) {
-                break;
-            }
-
-            if (trigger->type == TT_SWITCH && trigger->timer && switch_off) {
-                break;
-            }
-
-            if (g_Camera.num == g_Camera.last && trigger->type != TT_SWITCH) {
-                break;
-            }
-
-            g_Camera.timer = cam_data->timer;
-            if (g_Camera.timer != 1) {
-                g_Camera.timer *= LOGIC_FPS;
-            }
-
-            if (cam_data->one_shot) {
-                camera->flags |= IF_ONE_SHOT;
-            }
-
-            g_Camera.speed = cam_data->glide + 1;
-            g_Camera.type = is_heavy ? CAM_HEAVY : CAM_FIXED;
-            break;
-        }
-
-        case TO_TARGET:
-            camera_item = Item_Get((int16_t)(intptr_t)cmd->parameter);
-            break;
-
-        case TO_SINK: {
-            const OBJECT_VECTOR *const sink =
-                Camera_GetFixedObject((int16_t)(intptr_t)cmd->parameter);
-
-            if (g_Lara.lot.required_box != sink->flags) {
-                g_Lara.lot.target = sink->pos;
-                g_Lara.lot.required_box = sink->flags;
-            }
-
-            g_Lara.current_active = sink->data * 6;
-            break;
-        }
-
-        case TO_FLIPMAP: {
-            const int16_t flip_slot = (int16_t)(intptr_t)cmd->parameter;
-            int32_t slot_flags = Room_GetFlipSlotFlags(flip_slot);
-            flip_available = true;
-            if (slot_flags & IF_ONE_SHOT) {
-                break;
-            }
-
-            if (trigger->type == TT_SWITCH) {
-                slot_flags ^= trigger->mask;
-            } else if (trigger->mask) {
-                slot_flags |= trigger->mask;
-            }
-
-            if ((slot_flags & IF_CODE_BITS) == IF_CODE_BITS) {
-                if (trigger->one_shot) {
-                    slot_flags |= IF_ONE_SHOT;
-                }
-
-                if (!flip_status) {
-                    flip_map = true;
-                }
-            } else if (flip_status) {
-                flip_map = true;
-            }
-
-            Room_SetFlipSlotFlags(flip_slot, slot_flags);
-            break;
-        }
-
-        case TO_FLIPON: {
-            const int16_t flip_slot = (int16_t)(intptr_t)cmd->parameter;
-            const int32_t slot_flags = Room_GetFlipSlotFlags(flip_slot);
-            flip_available = true;
-            if ((slot_flags & IF_CODE_BITS) == IF_CODE_BITS && !flip_status) {
-                flip_map = true;
-            }
-            break;
-        }
-
-        case TO_FLIPOFF: {
-            const int16_t flip_slot = (int16_t)(intptr_t)cmd->parameter;
-            const int32_t slot_flags = Room_GetFlipSlotFlags(flip_slot);
-            flip_available = true;
-            if ((slot_flags & IF_CODE_BITS) == IF_CODE_BITS && flip_status) {
-                flip_map = true;
-            }
-            break;
-        }
-
-        case TO_FLIPEFFECT:
-            new_effect = (int16_t)(intptr_t)cmd->parameter;
-            break;
-
-        case TO_FINISH:
-            Game_SetIsLevelComplete(true);
-            break;
-
-        case TO_CD:
-            M_TriggerMusicTrack((int16_t)(intptr_t)cmd->parameter, trigger);
-            break;
-
-        case TO_SECRET: {
-            const int16_t secret_num = 1 << (int16_t)(intptr_t)cmd->parameter;
-            if (Stats_AddSecret(secret_num)) {
-                Music_Play(MX_SECRET, MPM_ALWAYS);
-            }
-            break;
-        }
-        }
-    }
-
-    if (camera_item
-        && (g_Camera.type == CAM_FIXED || g_Camera.type == CAM_HEAVY)) {
-        g_Camera.item = camera_item;
-    }
-
-    if (flip_map) {
-        Room_FlipMap();
-    }
-
-    if (new_effect != -1 && (flip_map || !flip_available)) {
-        Room_SetFlipEffect(new_effect);
-        Room_SetFlipTimer(0);
     }
 }
